Change key access from symbol to string.
I'm guessing that because the data_bags' items are stored as JSON the
symbol lookup of members isn't working correctly.  I'm commiting this so
that I can test in a prepared environment (if we decide to rebase we
can).

Also, I've added info and debug messages that will help point at
the problems with this in the future.