Add rabbit MQ user recipe.
I've added a recipe that when included on a rabbit system will create
the expected queue user in rabbit for margarine.

This adds to our interface but doesn't break back backwards
compatibility so I've bumped the minor version with this commit.

This new recipe adds a suggestion (not a requirement for rabbitmq).
This is due to the recipe needing the rabbitmq cookbook but not being
pulled in by default.  It should only be called on rabbit systems and
not on margarine direct systems.