Move logging configuration to attributes.
The logging configuration needed to drop SysLogHandler configurations as
some pythons on Ubuntu are having a hard time with it.  If anyone starts
using this in the wild it will need to be reported as a bug to margarine
so we can determine why the error is occurring.  Otherwise, it's pretty
much the same interface just attributes rather than data bags.